// src/api/model/apiResponsePageResponsePlatformListItemResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 주뜨 청년부 앱 API
 * 주뜨 청년부 앱 REST API 명세
 * OpenAPI spec version: v1
 */
import type { PageResponsePlatformListItemResponse } from './pageResponsePlatformListItemResponse';

export interface ApiResponsePageResponsePlatformListItemResponse {
  status?: number;
  message?: string;
  data?: PageResponsePlatformListItemResponse;
}