// src/api/model/apiResponsePlatformDetailResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 주뜨 청년부 앱 API
 * 주뜨 청년부 앱 REST API 명세
 * OpenAPI spec version: v1
 */
import type { PlatformDetailResponse } from './platformDetailResponse';

export interface ApiResponsePlatformDetailResponse {
  status?: number;
  message?: string;
  data?: PlatformDetailResponse;
}