// src/api/model/platformDetailResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 주뜨 청년부 앱 API
 * 주뜨 청년부 앱 REST API 명세
 * OpenAPI spec version: v1
 */
import type { PlatformDetailResponseApprovalStatus } from './platformDetailResponseApprovalStatus';
import type { PlatformDetailResponseClosedStatus } from './platformDetailResponseClosedStatus';

export interface PlatformDetailResponse {
  platformId?: number;
  title?: string;
  scheduleText?: string;
  startsAt?: string;
  endsAt?: string;
  location?: string;
  content?: string;
  purpose?: string;
  etc?: string;
  posterUrl?: string;
  approvalStatus?: PlatformDetailResponseApprovalStatus;
  recruiting?: boolean;
  operating?: boolean;
  closedStatus?: PlatformDetailResponseClosedStatus;
  ownerMemberId?: number;
  ownerName?: string;
  approvedMemberCount?: number;
  createdAt?: string;
  updatedAt?: string;
}