// src/api/model/platformListItemResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 주뜨 청년부 앱 API
 * 주뜨 청년부 앱 REST API 명세
 * OpenAPI spec version: v1
 */
import type { PlatformListItemResponseApprovalStatus } from './platformListItemResponseApprovalStatus';
import type { PlatformListItemResponseClosedStatus } from './platformListItemResponseClosedStatus';

export interface PlatformListItemResponse {
  platformId?: number;
  title?: string;
  scheduleText?: string;
  startsAt?: string;
  endsAt?: string;
  location?: string;
  posterUrl?: string;
  approvalStatus?: PlatformListItemResponseApprovalStatus;
  recruiting?: boolean;
  operating?: boolean;
  closedStatus?: PlatformListItemResponseClosedStatus;
  ownerMemberId?: number;
  ownerName?: string;
  approvedMemberCount?: number;
  createdAt?: string;
}